backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.5
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
